requirements.txt
Django==1.8.4
django-autoslug==1.8.0
django-taggit==0.17.0
